AuthViews: extracted style + layout changes
