+++ src/views/ResetView/index.js
@@ -6,22 +6,7 @@ import Button from '@material-ui/core/Button';
 import TextField from '@material-ui/core/TextField';
 import Typography from '@material-ui/core/Typography';
 
-
-const styles = theme => ({
-  actions: {
-    marginTop: '16px',
-    textAlign: 'center'
-  },
-  bold: {
-    fontWeight: theme.typography.fontWeightMedium,
-  },
-  submit: {
-    display: 'block',
-    marginLeft: 'auto',
-    marginRight: 'auto',
-    marginTop: '16px'
-  }
-});
+import authStyle from '../../assets/jss/authStyle';
 
 
 class ResetForm extends React.Component {
@@ -78,7 +63,7 @@ class ResetForm extends React.Component {
                    </Typography>
                  </React.Fragment>
              }
-             <div className={classes.actions}>
+             <div className={classes.action}>
                <Button
                  color="primary"
                  size="small"
@@ -95,4 +80,4 @@ ResetForm.propTypes = {
   classes: PropTypes.object.isRequired,
 };
 
-export default withStyles(styles)(ResetForm);
+export default withStyles(authStyle)(ResetForm);
